Backend/fastapi_app/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
sqlalchemy==2.0.28
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1
python-multipart==0.0.6
isodate==0.6.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
jinja2>=3.1.2
python-jose>=3.3.0
passlib>=1.7.4
google-api-python-client>=2.100.0

# Machine learning libraries
xgboost>=1.5.0
lightgbm>=3.3.2
catboost>=1.0.0
joblib>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.2
openpyxl>=3.0.9 